Reduce complexity at top level control flow logic and move logic down #refactoring
